Lots of updates
I deleted a bunch of .java files that were redundant. I simplified the whole thing into a JFrame and two JPanels. One for the headlines, and one for the four buttons on the bottom (different news sources)
The next step is to have each button respond by changing the list of news on top